# docs/admin/admin_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "接口",
        "title": "管理系统API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/admin/server-control/v1/audit-events": {
            "get": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Server Control"
                ],
                "summary": "List server-control audit events",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (maximum 100)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/admin.ControlAPIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/admin.ControlAuditEventList"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/admin.ControlAPIProblem"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/admin.ControlAPIProblem"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/admin.ControlAPIProblem"
                        }
                    }
                }
            }
        },
        "/admin/server-control/v1/instances": {
            "get": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Server Control"
                ],
                "summary": "List configured Starry instances",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/admin.ControlAPIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/service.ServerControlInstance"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/admin.ControlAPIProblem"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/admin.ControlAPIProblem"
                        }
                    }
                }
            }
        },
        "/admin/server-control/v1/instances/{id}/allocation-simulations": {
            "post": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Server Control"
                ],
                "summary": "Simulate Relay allocation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Deployment instance ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Typed simulation input",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/starrycontrol.SimulationInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/admin.ControlAPIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/starrycontrol.SimulationResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/admin.ControlAPIProblem"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/admin.ControlAPIProblem"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/admin.ControlAPIProblem"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/admin.ControlAPIProblem"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/admin.ControlAPIProblem"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/admin.ControlAPIProblem"
                        }
                    }
                }
            }
        },
        "/admin/server-control/v1/instances/{id}/capabilities": {
            "get": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Server Control"
                ],
                "summary": "Read Starry capabilities",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Deployment instance ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/admin.ControlAPIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/starrycontrol.Capabilities"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/admin.ControlAPIProblem"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/admin.ControlAPIProblem"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/admin.ControlAPIProblem"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/admin.ControlAPIProblem"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/admin.ControlAPIProblem"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "admin.ControlAPIProblem": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/admin.ControlAPIError"
                }
            }
        },
        "admin.ControlAPIResponse": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
        "admin.ControlAuditEventList": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AdminAuditEvent"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "service.ServerControlInstance": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean"
                },
                "enabled": {
                    "type": "boolean"
                },
                "error_code": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "read_only": {
                    "type": "boolean"
                }
            }
        },
...
        "starrycontrol.Capabilities": {
            "type": "object",
            "properties": {
                "capabilities": {
                    "$ref": "#/definitions/starrycontrol.CapabilityVersions"
                },
                "config": {
                    "$ref": "#/definitions/starrycontrol.ConfigCapabilities"
                },
                "instance": {
                    "$ref": "#/definitions/starrycontrol.InstanceInfo"
                },
                "limits": {
                    "$ref": "#/definitions/starrycontrol.AgentLimits"
                },
                "protocol": {
                    "$ref": "#/definitions/starrycontrol.ProtocolInfo"
                }
            }
        },
...
        "starrycontrol.SimulationInput": {
            "type": "object",
            "properties": {
                "client_a": {
                    "$ref": "#/definitions/starrycontrol.SimulationClient"
                },
                "client_b": {
                    "$ref": "#/definitions/starrycontrol.SimulationClient"
                },
                "expected_config_generation": {
                    "type": "integer"
                },
                "explain": {
                    "type": "boolean"
                },
                "transport": {
                    "$ref": "#/definitions/starrycontrol.Transport"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "token": {
            "type": "apiKey",
            "name": "api-token",
            "in": "header"
        }
    }
}